+++ packages/ag-charts-types/src/chartBuilderOptions.ts
@@ -1,7 +1,7 @@
 /* eslint-disable sonarjs/class-name */
 import type { AgInitialStateOptions } from './api/initialStateOptions';
 import type { AgBaseChartOptions } from './chart/chartOptions';
-import type { AgBaseChartThemeOptions, AgBaseGaugePresetThemeOptions } from './chart/themeOptions';
+import type { AgBaseChartThemeOptions, AgBaseGaugePresetThemeOptions } from './chart/themableOptions';
 import type { PixelSize } from './chart/types';
 import type { AgFinancialChartPresets } from './presets/financial/financialOptions';
 import type { AgGaugePresets } from './presets/gauge/gaugeOptions';
